move aoss to driver
