+++ graphiti_core/driver/neo4j_driver.py
@@ -18,7 +18,9 @@ import logging
 from collections.abc import Coroutine
 from typing import Any
 
+import boto3
 from neo4j import AsyncGraphDatabase, EagerResult
+from opensearchpy import OpenSearch, Urllib3AWSV4SignerAuth, Urllib3HttpConnection
 from typing_extensions import LiteralString
 
 from graphiti_core.driver.driver import GraphDriver, GraphDriverSession, GraphProvider
@@ -29,13 +31,35 @@ logger = logging.getLogger(__name__)
 class Neo4jDriver(GraphDriver):
     provider = GraphProvider.NEO4J
 
-    def __init__(self, uri: str, user: str | None, password: str | None, database: str = 'neo4j'):
+    def __init__(
+        self,
+        uri: str,
+        user: str | None,
+        password: str | None,
+        database: str = 'neo4j',
+        aoss_host: str | None = None,
+        aoss_port: int | None = None,
+    ):
         super().__init__()
         self.client = AsyncGraphDatabase.driver(
             uri=uri,
             auth=(user or '', password or ''),
         )
         self._database = database
+
+        self.aoss_client = None
+        if aoss_host and aoss_port:
+            session = boto3.Session()
+            self.aoss_client = OpenSearch(
+                hosts=[{'host': aoss_host, 'port': aoss_port}],
+                http_auth=Urllib3AWSV4SignerAuth(
+                    session.get_credentials(), session.region_name, 'aoss'
+                ),
+                use_ssl=True,
+                verify_certs=True,
+                connection_class=Urllib3HttpConnection,
+                pool_maxsize=20,
+            )
 
     async def execute_query(self, cypher_query_: LiteralString, **kwargs: Any) -> EagerResult:
         # Check if database_ is provided in kwargs.
@@ -61,6 +85,8 @@ class Neo4jDriver(GraphDriver):
         return await self.client.close()
 
     def delete_all_indexes(self) -> Coroutine[Any, Any, EagerResult]:
+        if self.aoss_client:
+            self.delete_all_indexes_impl()
         return self.client.execute_query(
             'CALL db.indexes() YIELD name DROP INDEX name',
         )
